requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==3.1.14
djangorestframework==3.14.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2019.3
sqlparse==0.3.1
